+++ example/app/view/account/Main.js
@@ -2,8 +2,10 @@ Ext.define("StackMobSenchaTouchDemo.view.account.Main", {
     extend: 'Ext.Panel',
 
     requires: [
+        'StackMobSenchaTouchDemo.view.account.CreateUser',
         'StackMobSenchaTouchDemo.view.account.Login',
-        'StackMobSenchaTouchDemo.view.account.Profile'
+        'StackMobSenchaTouchDemo.view.account.Profile',
+        'StackMobSenchaTouchDemo.view.account.ResetPassword'
     ],
 
     id: 'accountMainView',
@@ -16,6 +18,10 @@ Ext.define("StackMobSenchaTouchDemo.view.account.Main", {
             xclass: 'StackMobSenchaTouchDemo.view.account.Login'
         },{
             xclass: 'StackMobSenchaTouchDemo.view.account.Profile'
+        },{
+            xclass: 'StackMobSenchaTouchDemo.view.account.ResetPassword'
+        },{
+            xclass: 'StackMobSenchaTouchDemo.view.account.CreateUser'
         }]
     }
 });